move the PdbScanState struct and code into its own module
