+++ pg_search/src/postgres/customscan/pdbscan/scan_state.rs
@@ -1,0 +1,123 @@
+use crate::index::state::{SearchResults, SearchState};
+use crate::postgres::customscan::builders::custom_state::CustomScanStateWrapper;
+use crate::postgres::customscan::pdbscan::projections::snippet::SnippetInfo;
+use crate::postgres::customscan::pdbscan::PdbScan;
+use crate::postgres::customscan::CustomScanState;
+use crate::postgres::utils::VisibilityChecker;
+use crate::schema::SearchConfig;
+use pgrx::{name_data_to_str, pg_sys};
+use std::collections::HashMap;
+use tantivy::snippet::SnippetGenerator;
+
+const SORT_ASCENDING: u32 = pg_sys::BTLessStrategyNumber;
+const SORT_DESCENDING: u32 = pg_sys::BTGreaterStrategyNumber;
+
+#[derive(Debug, Default, Copy, Clone)]
+#[repr(u32)]
+pub enum SortDirection {
+    #[default]
+    Asc = pg_sys::BTLessStrategyNumber,
+    Desc = pg_sys::BTGreaterStrategyNumber,
+}
+
+impl From<SortDirection> for crate::index::state::SortDirection {
+    fn from(value: SortDirection) -> Self {
+        match value {
+            SortDirection::Asc => crate::index::state::SortDirection::Asc,
+            SortDirection::Desc => crate::index::state::SortDirection::Desc,
+        }
+    }
+}
+
+impl From<i32> for SortDirection {
+    fn from(value: i32) -> Self {
+        SortDirection::from(value as u32)
+    }
+}
+
+impl From<u32> for SortDirection {
+    fn from(value: u32) -> Self {
+        match value {
+            pg_sys::BTLessStrategyNumber => SortDirection::Asc,
+            pg_sys::BTGreaterStrategyNumber => SortDirection::Desc,
+            _ => panic!("unrecognized sort strategy number: {value}"),
+        }
+    }
+}
+
+impl From<SortDirection> for u32 {
+    fn from(value: SortDirection) -> Self {
+        value as _
+    }
+}
+
+#[derive(Default)]
+pub struct PdbScanState {
+    pub heaprelid: pg_sys::Oid,
+    pub indexrelid: pg_sys::Oid,
+    pub rti: pg_sys::Index,
+
+    pub index_name: String,
+    pub index_uuid: String,
+    pub key_field: String,
+    pub search_state: Option<SearchState>,
+    pub search_config: SearchConfig,
+    pub search_results: SearchResults,
+
+    pub limit: Option<usize>,
+    pub sort_direction: Option<SortDirection>,
+
+    pub heaprel: Option<pg_sys::Relation>,
+    pub indexrel: Option<pg_sys::Relation>,
+    pub lockmode: pg_sys::LOCKMODE,
+
+    pub snapshot: Option<pg_sys::Snapshot>,
+    pub visibility_checker: Option<VisibilityChecker>,
+
+    pub need_scores: bool,
+    pub snippet_generators: HashMap<SnippetInfo, Option<SnippetGenerator>>,
+    pub score_funcoid: pg_sys::Oid,
+    pub snippet_funcoid: pg_sys::Oid,
+    pub var_attname_lookup: HashMap<(i32, pg_sys::AttrNumber), String>,
+
+    pub scan_func: Option<fn(&mut CustomScanStateWrapper<PdbScan>) -> *mut pg_sys::TupleTableSlot>,
+}
+
+impl CustomScanState for PdbScanState {}
+
+impl PdbScanState {
+    #[inline(always)]
+    pub fn need_scores(&self) -> bool {
+        self.need_scores
+    }
+
+    #[inline(always)]
+    pub fn need_snippets(&self) -> bool {
+        !self.snippet_generators.is_empty()
+    }
+
+    #[inline(always)]
+    pub fn snapshot(&self) -> pg_sys::Snapshot {
+        self.snapshot.unwrap()
+    }
+
+    #[inline(always)]
+    pub fn heaprel(&self) -> pg_sys::Relation {
+        self.heaprel.unwrap()
+    }
+
+    #[inline(always)]
+    pub fn heaprelname(&self) -> &str {
+        unsafe { name_data_to_str(&(*(*self.heaprel()).rd_rel).relname) }
+    }
+
+    #[inline(always)]
+    pub fn heaptupdesc(&self) -> pg_sys::TupleDesc {
+        unsafe { (*self.heaprel()).rd_att }
+    }
+
+    #[inline(always)]
+    pub fn visibility_checker(&mut self) -> &mut VisibilityChecker {
+        self.visibility_checker.as_mut().unwrap()
+    }
+}
